telegram_uploader_bot/requirements.txt
telethon
python-dotenv
requests
cryptg
hachoir
